+++ config/utils/applyOpacityValue.ts
@@ -1,8 +1,3 @@
 export const applyOpacityValue = (variable: string) => {
-  return ({ opacityValue }: { opacityValue?: number }) => {
-    if (opacityValue === undefined) {
-      return `rgb(var(${variable}))`;
-    }
-    return `rgb(var(${variable}) / ${opacityValue})`;
-  };
+  return `rgb(var(${variable}) / <alpha-value>)`;
 };