+++ config/utils/safeList.ts
@@ -1,9 +1,12 @@
-import { SafelistConfig } from "tailwindcss/types/config";
+type SafeListEntry = {
+  pattern: RegExp;
+  variants?: string[];
+};
 
-export const safeList: Partial<SafelistConfig> = [
+export const safeList: SafeListEntry[] = [
   {
     pattern:
-      /^(bg|to|via|from|text|fill|stroke|border|outline)-((primary|secondary|warning|error|success|content1|content2|content3|backgroundPrimary|backgroundSecondary|border|black|white|whiteInverted|blackInverted)|((blue|red|green|purple|pink|gray|cyan|slate)-(1|2|3|4|5|6|7|8|9|10)\D))/,
+      /^(bg|to|via|from|text|fill|stroke|border|outline)-((primary|secondary|warning|error|success|content1|content2|content3|backgroundPrimary|backgroundSecondary|border|black|white|whiteInverted|blackInverted)|((blue|red|green|purple|pink|gray|cyan|slate)-(1|2|3|4|5|6|7|8|9|10)(?!\d)))/,
     variants: [
       // "first",
       // "last",