+++ docs/components/styles/Community.tsx
@@ -39,7 +39,7 @@ export const Community = () => {
         </Link>
         <Link
           aria-label="RippleUI Github"
-          href={"https://github.com/Siumauricio/rippleui"}
+          href={"https://github.com/wisdom-plus/NamiUI"}
           target="_blank"
           className="card cursor-pointer p-8  shadow-md transition-colors hover:bg-gray-4"
         >