+++ docs/components/styles/Navbar.tsx
@@ -102,7 +102,7 @@ export const Navbar = () => {
           </Link>
           <Link
             aria-label="Github"
-            href={"https://github.com/Siumauricio/rippleui"}
+            href={"https://github.com/wisdom-plus/NamiUI"}
             target="_blank"
             className="hidden cursor-pointer lg:flex"
           >