+++ examples/nextjs-example/tailwind.config.js
@@ -7,5 +7,5 @@ module.exports = {
   theme: {
     extend: {},
   },
-  plugins: [require("rippleui")],
+  plugins: [require("namiui")],
 };